+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- [breaking] [d4lcontext]: `GetUserID`, `GetUserIDFromCtx` and `WithUserID` work with UUIDs instead of strings
+
 ### Deprecated
 
 ### Removed